put the interval class in its own module
